LayoutTests/fast/forms/datetimelocal/datetime-local-value-sanitization-expected.txt

Tests the sanitization of the value of  elements of type datetime-local.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.value is "2000-01-01T12:30:00.5"
PASS input.value is "2000-01-01T12:30:00.6"
PASS input.value is "2000-01-01T12:30:00.7"
PASS input.value is "2000-01-01T12:30:00.04"
PASS input.value is "2000-01-01T12:30:00.003"
PASS input.value is ""
PASS input.value is "2000-01-01T12:30"
PASS input.value is "2000-01-01T12:30"
PASS input.value is "2000-01-01T12:30:01"
PASS input.value is "2000-01-01T12:30:10"
PASS successfullyParsed is true

TEST COMPLETE
